Fix: configuration variable that overrides a template variable in env0_environment shows drift
